+++ packages/emission/src/scope_new/index.js
@@ -1,0 +1,11 @@
+//@flow
+
+export {Scope} from './scope'
+export {dispatch} from './dispatch'
+export {Emission} from './instance'
+
+import * as EventRef from './ref'
+import * as EventSet from './set'
+import * as EventWatcher from './watcher'
+
+export {EventRef, EventSet, EventWatcher}