+++ packages/store/src/__test__/index.test.js
@@ -1,0 +1,7 @@
+//@flow
+
+import {createStore} from '..'
+
+test('smoke', () => {
+ expect(createStore).toBeDefined()
+})